Minor Refactors: Rename classes and files Extract MainPage to separate dart file
